# backend/src/config/mobileOpenApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lidifin Mobile API",
    "version": "1.0.0",
    "description": "Dedicated mobile-facing API contract for native clients. All paths are relative to /api. Primary authentication is username/password via JWT access + refresh tokens."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative API base path"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Primary mobile authentication. Obtain with POST /auth/login and refresh with POST /auth/refresh."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Secondary device authentication for paired devices and integrations."
      },
      "streamTokenAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Optional query token for stream URLs when attaching Authorization headers is inconvenient."
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Web UI session authentication. Included for completeness."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": ["id", "username", "role"],
        "properties": {
          "id": { "type": "string" },
          "username": { "type": "string" },
          "role": {
            "type": "string",
            "enum": ["user", "admin"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "onboardingComplete": {
            "type": "boolean"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "details": {}
        }
      },
      "AuthTokens": {
        "type": "object",
        "required": ["token", "refreshToken", "user"],
        "properties": {
          "token": { "type": "string" },
          "refreshToken": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "Requires2FA": {
        "type": "object",
        "required": ["requires2FA", "message"],
        "properties": {
          "requires2FA": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "2FA token required" }
        }
      },
...
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login with username and password",
        "description": "Primary mobile authentication flow. Returns JWT access and refresh tokens. If two-factor auth is enabled and no second factor is provided, the response indicates that 2FA is required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string", "format": "password" },
                  "token": {
                    "type": "string",
                    "description": "Optional TOTP or recovery code when 2FA is enabled"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, or second-factor confirmation required",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/AuthTokens" },
                    { "$ref": "#/components/schemas/Requires2FA" }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh an access token",
        "description": "Exchanges a valid refresh token for a new access token and refresh token pair.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["token", "refreshToken"],
                  "properties": {
                    "token": { "type": "string" },
                    "refreshToken": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing refresh token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current authenticated user",
        "security": [
          { "bearerAuth": [] },
          { "apiKeyAuth": [] },
          { "sessionAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}